ai services/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# Embeddings (pure Python, no C++ needed)
sentence-transformers==3.2.1

# Anthropic LLM (optional — works without API key in mock mode)
anthropic==0.40.0

# Async Postgres
asyncpg==0.30.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
numpy==2.1.3
